more examples, exercise solutions

+++ examples/ex_twosumfloat.py
@@ -1,14 +1,14 @@
 # Read string from keyboard into variable x
 x = input("Enter a number: ")
 
-# Convert x to int and store it back in x again
-x = int(x)
+# Convert x to float and store it back in x again
+x = float(x)
 
 # Read string from keyboard into variable y
 y = input("Enter another number: ")
 
-# Convert y to int and store it back in y again
-y = int(y)
+# Convert y to float and store it back in y again
+y = float(y)
 
 # Print the sum of x + y
 print("The sum of the two numbers is:", x + y)
